Refactoring envelopes package to be more purely Immutable.
In the past, I relied on passing everything by value to make it immutable. However certain types, like maps pretty well destroy that line of thought.

Another thing I strove to achieve with this change is getting the correct balance of ordered vs non-ordered data structures. Marshaling needs to be
ridgid in its order, because the SHA1 IDs need to be reproducable and deterministic. Unmarshaling on the other hand, I'd like to soak up any bugs
that the client came up with and just keep moving. Often this means that I use maps instead of arrays.