`persist.Fetch` -> `[]byte` not `io.Reader`
Fetching returning a Reader and a Reader alone didn't offer a way to close the reader at the end of dealing with it. I could have chosen to use ReadCloser, but none of the objects should really be huge anyway. There are plenty of other places in this code base where any given object will be entirely loaded into memory. Engineering choices for the win.