Converting envelopes to no longer use the immutable pattern.
There were just too many times where I wanted to be able to mutate an object. The straw that broke the camel's back was when I realized that I was going to need to move the serialization/deserialization logic into the base package to allow for users to not need to understand the object model deeply.