Adding interface collections for Repositories
Until now, a lot of effort has been maintained to keep things as granular and modular as possible. This is in keeping with the idea that we should be building up lots of individual reusable components instead of having a top-down Repository object that just relies on oppotunistic dependency injection. In other languages, it could be that it's much more idiomatic to do the other way, but the same attention should be paid to keeping many many pieces reusable.