Adding opt. additionalParents to `persist.Commit`
The main value of the Commit function is its ability to lookup and
advance the current branch. When you're creating a merge commit, the
same problem exists, but additional parents get tacked on. For that
reason, it doesn't really make sense to have two separate functions.

Plus, the actual Merge process will need to include hooks for people to
manually resolve conflicts. So that will need to be a very different
function.