Removing Conflict Finding from Merge function
After thinking through it more deeply, a few things are clear to me:
1. Duplicate transactions, i.e. transactions that were independently
   added in multiple branches that are now being merged together, are
   the merge-conflict analog here.
2. Finding duplicate transactions will actually be sort of expensive,
   and it'll definitely be very complicated. There's a sort of basic
   premise I started with that it's just a set-interesection problem,
   and like it is, but then how do I handle various places where
   transactions were duplicates but then got revoked. Not to mention
   just run of the mill crazy branch structure scenarios.
3. I think there's a real chance that people may want different
   duplicate detection strategies at different times.

For now, I'm going to just take it out of this signature and keep
moving. I want to go actually use it for a while and see if it feels
like it's even necessary to do something complicated there. It could
always be a preliminary method that runs where all but one of the dupes
gets revoked, then this addded Merge function gets added.